requirements.txt
Flask
flask-cors
numpy
scikit-learn
pandas
matplotlib
tensorflow
firebase-admin
python-dotenv
Pyrebase
requests-toolbelt==0.10.1